+++ CHANGES.rst
@@ -6,6 +6,14 @@
  Unreleased
 ************
 
+**Added**
+
+- Added support for missing docutils roles and directives.
+
+**Fixed**
+
+- The rows of a ``csv-table`` directive are no longer merged into a single line.
+
 ********************
  2.2.1 (2026/08/28)
 ********************